refactor(frontend): Add lazy-loaded assignment module
