+++ frontend/src/app/assignment/modules/assignment/assignment/assignment.component.ts
@@ -1,13 +1,12 @@
-import {DOCUMENT} from '@angular/common';
-import {Component, Inject, OnInit} from '@angular/core';
+import {Component, OnInit} from '@angular/core';
 import {ActivatedRoute, Router} from '@angular/router';
 import {combineLatest} from 'rxjs';
 import {switchMap} from 'rxjs/operators';
-import {ToastService} from '../../../toast.service';
-import {assignmentChildRoutes} from '../../assignment-routing.module';
-import Assignment from '../../model/assignment';
-import {AssignmentService} from '../../services/assignment.service';
-import {SolutionService} from '../../services/solution.service';
+import {ToastService} from '../../../../toast.service';
+import Assignment from '../../../model/assignment';
+import {AssignmentService} from '../../../services/assignment.service';
+import {SolutionService} from '../../../services/solution.service';
+import {assignmentChildRoutes} from '../assignment-routing.module';
 
 @Component({
   selector: 'app-assignment',
@@ -17,7 +16,7 @@ import {SolutionService} from '../../services/solution.service';
 export class AssignmentComponent implements OnInit {
   assignment?: Assignment;
   importing = false;
-  routes = assignmentChildRoutes.filter(r => r.path !== 'token');
+  routes = assignmentChildRoutes;
 
   constructor(
     public route: ActivatedRoute,
