+++ frontend/src/app/assignment/modules/assignment/share/share.component.ts
