+++ frontend/src/app/assignment/modules/assignment/solution-table/solution-table.component.ts
