+++ frontend/src/app/assignment/modules/assignment/statistics/statistics.component.spec.ts
